ref(repos): Consolidate organizationRepositoriesV2 into organizationRepositories
Collapses the v2 directory into the canonical module and reorganizes
its internals:

- Moves all files from organizationRepositoriesV2/ into
  organizationRepositories/, deleting the v2 directory
- Organizes into hooks/ (useDeleteIntegration, useInstallationSettings,
  useRepoSearch, useSyncRepositories,
  organizationConfigIntegrationsQueryOptions) and components/
  (ScmRepositoryTable, ConnectProviderDropdown, NoIntegrationsEmptyState)
- Extracts ScmInstallation, ScmRepoMatches, and InstallationWrapperProps
  into a types.ts file
- Renames providerConfigLink.tsx → getProviderConfigUrl.ts now that the
  React component is gone
- Renames OrganizationRepositoriesV2 → OrganizationRepositories and
  makes it the default export
