+++ static/app/views/settings/organizationRepositories/components/scmRepositoryTable.spec.tsx
@@ -5,8 +5,8 @@ import {RepositoryFixture} from 'sentry-fixture/repository';
 import {render, screen, userEvent, within} from 'sentry-test/reactTestingLibrary';
 
 import type {Repository} from 'sentry/types/integrations';
+import type {ScmInstallation} from 'sentry/views/settings/organizationRepositories/types';
 
-import type {ScmInstallation} from './scmRepositoryTable';
 import {InstallationOverrideProvider, ScmRepositoryTable} from './scmRepositoryTable';
 
 // `useVirtualizer` only renders rows whose computed bounding rect overlaps
